Creation of frontend projects

+++ Backend/KlimabewacherApi/Models/Medicion.cs
@@ -5,8 +5,8 @@ namespace KlimabewacherApi.Properties
     {
         public int Id { get; set; }
         public string DeviceId { get; set; } = string.Empty;
-        public float Temperatura { get; set; }
-        public float Humedad { get; set; }
+        public double Temperatura { get; set; }
+        public double Humedad { get; set; }
         public string? Ubicacion { get; set; }
         public DateTime FechaHora { get; set; } = DateTime.Now;
     }
